% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fibo}
\alias{fibo}
\title{fibo}
\usage{
fibo(n)
}
\arguments{
\item{n}{the number of samples}
}
\value{
fibo sampler using Rcpp \code{n}
}
\description{
fibo sampler using Rcpp
}
\examples{
\dontrun{
a <- fibo(5)
}
}
