% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIS.R
\name{ISIS}
\alias{ISIS}
\title{Title}
\usage{
ISIS(x, y, d, xtrue)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{d}{d}

\item{xtrue}{xtrue}
}
\value{
listdata
}
\description{
Title
}
\examples{
\dontrun{
a=ISIS(x,y,d,xtrue)
}
}
