% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIS.R
\name{Lasso}
\alias{Lasso}
\title{About sure independent screening methods}
\usage{
Lasso(x, y, xtrue)
}
\arguments{
\item{x}{The design matrix, of dimensions n * p, without an intercept.}

\item{y}{The response vector of dimension n * 1.}

\item{xtrue}{The size of true model}
}
\value{
Returns an object with \item{ISIS_flag}{A sign of whether ISIS contains a true value.}
\item{AA}{Variables for ISIS screening.}
}
\description{
SIS and ISIS using R
}
