% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppc}
\alias{rcppc}
\title{rcppc is isprime}
\usage{
rcppc(x)
}
\arguments{
\item{x}{the number of samples}
}
\value{
is or not rpime using Rcpp \code{n}
}
\description{
isprime sampler using Rcpp
}
\examples{
\dontrun{
b <- rcppc(5)
}
}
